Youtube_notes_generator/pyproject.toml
[project]
name = "youtube-notes-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.27.0",
    "streamlit>=1.47.1",
]
